% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BiGraph.R
\docType{data}
\name{BiGraph}
\alias{BiGraph}
\title{bipartite graph}
\format{\preformatted{Class 'R6ClassGenerator' <BiGraph> object generator
  Inherits from: <Graph>
  Public:
    n1: NULL
    n2: NULL
    k: NULL
    is_adj: NULL
    set_graph_inc: function (graph_inc) 
    initialize: function (type = c("bipartite_regular"), n1, n2 = NULL, k, directed = TRUE, 
    get_graph: function (is_adj = TRUE) 
    as_adj: function () 
    as_inc: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "BiGraph_generator"
}}
\usage{
BiGraph
}
\description{
bipartite graph
}
\examples{
BiGraph$new(type = 'bipartite_regular', n1 = 50, k = 5, directed = T, is_adj = T)
}
\keyword{datasets}

