% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphs_algorithms.R
\name{bigraphs_rewiring}
\alias{bigraphs_rewiring}
\title{Generate bipartite graphs with different degree heterogeneity by rewiring links (or Nestedness optimization algrithm).}
\usage{
bigraphs_rewiring(graph)
}
\description{
1) start from a regular bipartite graph.
2) repeat rewiring a link to a new species who has more neighbors than the old species
until failed for enough times.
3) return to step 2)
}
\examples{
biGraph <- BiGraph$new(type = 'bipartite_regular', n1 = 10, k = 4, is_adj = FALSE)
graphs <- bigraphs_rewiring(biGraph$get_graph(is_adj = FALSE))
}

