% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphs_algorithms.R
\name{bigraphs_rewiring_onestep}
\alias{bigraphs_rewiring_onestep}
\title{rewiring a link to some node with more links. (richer get richer)}
\usage{
bigraphs_rewiring_onestep(B, connected = TRUE, ntry = 100)
}
\arguments{
\item{B}{incidence matrix of bipartite network, rows and cols represent two groups of nodes/species}

\item{connected,}{if the new graph should be connected?}

\item{ntry,}{how many to try?}
}
\value{
the incidence matrix whose nestedness has been optimized by rewiring a link.
}
\description{
rewiring a link to some node with more links. (richer get richer)
}
\details{
.
}

