% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_semicircle_bipartite_regular}
\alias{estimate_semicircle_bipartite_regular}
\title{estimate semicircle eigenvalue of a bipartite regular graph using alon-method, plus-twins-method, or an empirical method}
\usage{
estimate_semicircle_bipartite_regular(n, km)
}
\arguments{
\item{n,}{node number}

\item{km,}{node degree}
}
\description{
estimate semicircle eigenvalue of a bipartite regular graph using alon-method, plus-twins-method, or an empirical method
}

