% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate ggplot2 default color palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
Emulate ggplot2 default color palette
}
\references{
http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}

