% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{ggplot_timeseries}
\alias{ggplot_timeseries}
\title{ggplot a multivariate timeseries}
\usage{
ggplot_timeseries(timeseries, title = NULL, xlab = NULL, ylab = NULL,
  is_legend = FALSE, size = 0.5)
}
\arguments{
\item{timeseries}{a numeric matrix representing a multivariate timeseries, the first colnum is the index data}

\item{title,}{is_legend}
}
\description{
ggplot a multivariate timeseries
}

