% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{inc_to_adj}
\alias{inc_to_adj}
\title{transfer an incidence matrix to an adjacency matrix}
\usage{
inc_to_adj(inc)
}
\arguments{
\item{inc,}{an incidency matrix}
}
\value{
adj, an adiacency matrix
}
\description{
transfer an incidence matrix to an adjacency matrix
}

