% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimLV2Mutual.R
\name{jacfunc_lv2_cm}
\alias{jacfunc_lv2_cm}
\title{Jacobian function for model \code{model_lv2_cm}}
\usage{
jacfunc_lv2_cm(time, init, params)
}
\description{
Jacobian function for model \code{model_lv2_cm}
}

