% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MeanFieldMutual.R
\docType{data}
\name{MeanFieldMutual}
\alias{MeanFieldMutual}
\title{A Mean Field Model for Mutualistic ecological communities}
\format{\preformatted{Class 'R6ClassGenerator' <MeanFieldMutual> object generator
  Inherits from: <MutualModel>
  Public:
    rho: NULL
    Delta_real: NULL
    Delta_est: NULL
    alpha_real: NULL
    alpha_est: NULL
    rmin: NULL
    extension: 1.1
    xinit_sd: 0.1
    get_params: function () 
    update: function (s = NULL, c = NULL, km = NULL, m = NULL, h = NULL) 
    initialize: function (n1 = 20, s = 1, c = 0.01, km = 5, m = 1, h = 0.5) 
    sim_slv2: function (rmax = 0, steps = 1000, stepwise = 1, sigma = 0.02, 
    get_xstars: function (r, s, c, kc, m, km, h) 
    get_dot_semicircle: function (r, s, c, kc, m, km, h, n, graphm) 
  Private:
    get_rho: function (s, c, kc, m, km) 
    get_m_semicircle_real: function (graphm) 
    get_m_semicircle_est: function (n, km) 
    get_Delta_real: function (c, kc, m, km, graphm) 
    get_Delta_est: function (c, kc, m, km, n) 
    get_dot: function (r, rho, h) 
    get_rmin: function (s, c, kc, m, km, h) 
    get_m_tilde: function (r, s, c, kc, m, km, h) 
    update_inner: function () 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "MeanFieldMutual_generator"
}}
\usage{
MeanFieldMutual
}
\description{
A Mean Field Model for Mutualistic ecological communities
}
\examples{
meanfieldMutual <- MeanFieldMutual$new(n1 = 20, s = 1, c = 0.01, km = 5, m = 1, h = 0.5)
meanfieldMutual$sim_slv2()
}
\keyword{datasets}

