% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimLV2Mutual.R
\name{model_lv2_cm}
\alias{model_lv2_cm}
\title{Lotka-Volterra (LV) Equations of Holling type II for a community mixed by Competition and Mutualism interactions}
\usage{
model_lv2_cm(time, init, params, ...)
}
\arguments{
\item{time,}{time steps of simulation}

\item{init,}{the initial state of the LV system, a vector}

\item{params,}{parameters passed to LV model, a list of:
\describe{
  \item{r}{a vector of the intrinsic growth rates of species}
  \item{C}{a matrix of intra-species and inter-species competitions}
  \item{M}{a matrix of mutualism interactions among species}
  \item{h}{the saturate coefficient, handling time of species feed}
}}
}
\value{
the derivation
}
\description{
Lotka-Volterra (LV) Equations of Holling type II for a community mixed by Competition and Mutualism interactions
}

