% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MutualModel.R
\docType{data}
\name{MutualModel}
\alias{MutualModel}
\title{The root class of Model for Mutualistic systems.}
\format{\preformatted{Class 'R6ClassGenerator' <MutualModel> object generator
  Public:
    n1: NULL
    n2: NULL
    n: NULL
    s: NULL
    kc: NULL
    c: NULL
    km: NULL
    m: NULL
    h: NULL
    graphc: NULL
    graphm: NULL
    simSLV2: NULL
    simLV2: NULL
    state_slv2: FALSE
    state_lv2: FALSE
    initialize: function (n1 = 20, s = 1, c = 0.01, km = 5, m = 1, h = 0.5) 
    sim_slv2: function (rmax, delta, steps, stepwise, sigma, xinit) 
    sim_lv2: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "MutualModel_generator"
}}
\usage{
MutualModel
}
\description{
This class has restrictions. First, species have to be splitted to two groups, mutualistic interactions happen between groups while competitive interactions happen within groups, thus, the graph of mutualistic interactions is bipartite. Second, the number of species of two groups are same. If these restrictions need to be released, we need SUPER class of this class.
}
\section{Fields}{

\describe{
\item{\code{n1}}{species number of group 1}

\item{\code{n2}}{species number of group 2}
}}
\keyword{datasets}

