% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric_derivative}
\alias{numeric_derivative}
\title{numerical differentiation algorithm to estimate the derivative of a function based on numerical data}
\usage{
numeric_derivative(data, h)
}
\description{
numerical differentiation algorithm to estimate the derivative of a function based on numerical data
}
\references{
https://en.wikipedia.org/wiki/Numerical_differentiation
}

