% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimLV2Mutual.R
\name{params_lv2_cm}
\alias{params_lv2_cm}
\title{parameters for model \code{\link{model_lv2_cm}}}
\usage{
params_lv2_cm(coeffs, graphc, graphm)
}
\arguments{
\item{coeffs}{a list of coefficients : list(n1 = n1, n2 = n2, r.row.mu = r, r.row.sd = r.sd, r.col.mu = r, r.col.sd = r.sd, s.mu = s, s.sd = s.sd, c.mu = c, c.sd = c.sd, m.mu = m, m.sd = m.sd, h.mu = h, h.sd = h.sd, delta = delta)}

\item{graphc,}{the adjacency matrix of competitive interactions}

\item{graphm,}{the adjacency matrix of mutualistic interactions}
}
\value{
a list of parameters for model \code{\link{model_lv2_cm}}
}
\description{
assign parameters for model \code{\link{model_lv2_cm}} according to a structural network and a couple of coefficients
}

