% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{preproc_trim_negative}
\alias{preproc_trim_negative}
\title{preprocess a multivariate timeseries}
\usage{
preproc_trim_negative(timeseries, trim_or_replace = c("trim", "replace"))
}
\description{
trim all data starting from the first negative value, or replace with 0
}

