% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_MeanFieldMutual}
\alias{print_MeanFieldMutual}
\title{visualize the output of MeanFieldMutual object}
\usage{
print_MeanFieldMutual(meanfieldMutual, remove.least = FALSE)
}
\arguments{
\item{meanfieldMutual}{a MeanFieldMutual object}
}
\description{
visualize the output of MeanFieldMutual object
}
\note{
can use a generic function \code{print}
}

