% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{run_stats_multivariate}
\alias{run_stats_multivariate}
\title{get covariances within moving windows along a multivariate timeseries}
\usage{
run_stats_multivariate(timeseries, winsize = 50, type = c("cov", "cor",
  "relative"))
}
\arguments{
\item{timeseries}{a numeric matrix of multivariate timeseries}

\item{winsize}{window size expressed as percentage of the timeseries length (must be numeric between 0 and 100). Default is 50.}

\item{type}{c('cov', 'cor', 'relative')}
}
\value{
a list of covariances measurements
}
\description{
get covariances within moving windows along a multivariate timeseries
}

