% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{runif2}
\alias{runif2}
\title{another form of uniform distribution between [mean - sd, mean + sd]}
\usage{
runif2(n, mean, sd)
}
\description{
another form of uniform distribution between [mean - sd, mean + sd]
}

