% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimObj.R
\docType{data}
\name{SimObj}
\alias{SimObj}
\title{The root class(interface) for all simulation objects}
\format{\preformatted{Class 'R6ClassGenerator' <SimObj> object generator
  Public:
    refObj: NULL
    state: NULL
    steps: NULL
    stepwise: NULL
    xinit: NULL
    params: NULL
    set_model: function () 
    set_times: function () 
    set_init: function (xinit) 
    set_params: function () 
    sim: function () 
    get_out: function () 
    initialize: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "SimObj_generator"
}}
\usage{
SimObj
}
\description{
A classical simulation includes several steps:
1. set model, such as Ordinary Differential Equaitons, Stochastic Differential Equation, or some kind of agent-based model.
2. set timesteps, [steps] with [stepwise] increment.
3. set initial values, for all the state variables.
4. set parameters, which adhere to the model.
5. start simulate, which will produce the output data.
6. get output of the simulation, in order for further data analysis.
}
\keyword{datasets}

