% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimYuima.R
\docType{data}
\name{SimYuima}
\alias{SimYuima}
\title{Simulation for (Multivariate) Stochastic Process implemented by \code{Yuima} package inherited from Interface \code{SimObj}}
\format{\preformatted{Class 'R6ClassGenerator' <SimYuima> object generator
  Inherits from: <SimObj>
  Public:
    drift: NULL
    diffusion: NULL
    variables: NULL
    initialize: function () 
    set_drift: function () 
    set_diffusion: function () 
    set_variables: function () 
    set_model: function () 
    set_times: function (steps, stepwise) 
    sim: function () 
    get_out: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "SimYuima_generator"
}}
\usage{
SimYuima
}
\description{
Simulation for (Multivariate) Stochastic Process implemented by \code{Yuima} package inherited from Interface \code{SimObj}
}
\section{Fields}{

\describe{
\item{\code{of}}{a stochastic process:
\describe{
\item{drift}{the drift vector}
\item{diffusion}{the diffusion matrix}
\item{variables}{name of state variables}
}}
}}
\keyword{datasets}

