% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwoblocksGraph.R
\docType{data}
\name{TwoblocksGraph}
\alias{TwoblocksGraph}
\title{graph represented by two blocks matrix}
\format{\preformatted{Class 'R6ClassGenerator' <TwoblocksGraph> object generator
  Inherits from: <Graph>
  Public:
    n1: NULL
    n2: NULL
    k: NULL
    initialize: function (type = "two_blocks_regular", n1, n2, k) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:StabEco>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "TwoblocksGraph_generator"
}}
\usage{
TwoblocksGraph
}
\description{
graph represented by two blocks matrix
}
\examples{
TwoblocksGraph$new(type = 'two_blocks_regular', n1 = 50, n2 = 50, k = 5)
}
\keyword{datasets}

